Overwrite flag, refactor
Add overwrite flag, allowing use to specify they wish to overwrite
any file in the destination with the same name as the source file.

By default, if a file with the same name as the source exists in the
destination directory, a number will be appended to the file to make
the filename unique.

This commit includes some refactoring including removing extra debug
messages, adding inline and cold attributes to functions, and the addition
of the lazy_static crate for a static Aho-Corasick file extension matcher.